Backend/image-conversion-service/requirements.txt
Flask==2.1.0
Pillow==8.3.2